PollLoop的多路复用

+++ Project_TCPServer_Plus/src/connection/Connection.hpp
@@ -46,7 +46,7 @@ private:
     std::string toLower(const std::string& s)
     {
         std::string res=s;
-        for(int i=0;i<s.size();i++)
+        for(size_t i=0;i<s.size();i++)
         {
             res[i]=tolower(s[i]);
         }
